organize tab step components into folders
