+++ packages/nextjs/app/v3/_components/ChooseFinalDetails/index.tsx
@@ -1,0 +1,11 @@
+import { ChooseInfo } from "./ChooseInfo";
+import { ChooseTokenAmounts } from "./ChooseTokenAmounts";
+
+export const ChooseFinalDetails = () => {
+  return (
+    <div className="flex flex-col gap-5">
+      <ChooseTokenAmounts />
+      <ChooseInfo />
+    </div>
+  );
+};
