+++ packages/nextjs/app/v3/_components/ChooseType/index.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { PoolTypeButton } from "./PoolTypeButton";
 import { ArrowUpRightIcon } from "@heroicons/react/24/solid";
 import { usePoolCreationStore } from "~~/hooks/v3";
 import { type SupportedPoolTypes, poolTypeMap } from "~~/utils/constants";
@@ -25,7 +26,6 @@ export function ChooseType() {
             ))}
           </div>
         </div>
-
         <div>
           <div className="text-xl bg-base-100 rounded-xl p-5 border border-neutral h-32 flex flex-col justify-center">
             {poolType ? poolTypeMap[poolType].description : staringInstructions}
@@ -33,30 +33,6 @@ export function ChooseType() {
         </div>
       </div>
     </>
-  );
-}
-
-function PoolTypeButton({ selectedPoolType }: { selectedPoolType: SupportedPoolTypes }) {
-  const { poolType, updatePool, tokenConfigs } = usePoolCreationStore();
-
-  const maxNumberOfTokens = poolType ? poolTypeMap[selectedPoolType].maxTokens : 0;
-
-  function handlePoolTypeSelection() {
-    console.log("handlePoolTypeSelection", selectedPoolType, maxNumberOfTokens);
-    updatePool({ poolType: selectedPoolType, tokenConfigs: tokenConfigs.slice(0, maxNumberOfTokens) });
-  }
-
-  return (
-    <button
-      className={`${
-        selectedPoolType === poolType ? `${selectedPoolStyles}` : `bg-base-100 ${hoverPoolStyles} shadow-lg`
-      } p-4 w-full rounded-xl`}
-      onClick={handlePoolTypeSelection}
-    >
-      <div className="flex flex-col text-center">
-        <div className="font-bold text-xl w-full">{poolTypeMap[selectedPoolType].label}</div>
-      </div>
-    </button>
   );
 }
 
@@ -74,9 +50,3 @@ const staringInstructions = (
     </a>
   </div>
 );
-
-const selectedPoolStyles =
-  "text-neutral-700 bg-gradient-to-r from-violet-300 via-violet-200 to-orange-300  [box-shadow:0_0_10px_5px_rgba(139,92,246,0.5)]";
-
-const hoverPoolStyles =
-  "hover:bg-gradient-to-r hover:from-violet-300 hover:via-violet-200 hover:to-orange-300 hover:text-neutral-700 hover:opacity-80";
